--- RSA_KEY_Generator/rsa_front.html
@@ -1,28 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-</head>
-<body>
-
-    <div>
-    Enter you Name: <input type="text">
-    Enter first prime number: <input type="text">
-    Enter second  prime number: <input type="text">
-    <button onclick="RSA()">Generate key</button>
-    </div>
-
-    <div id="e_key"> </div>
-    <div id="d_key">Your Decryption key is: </div>
-
-    <script>
-      function RSA(x,y)
-      {
-        let e= document.getElementById('e_key');
-        e.innerHTML="<h1> done</h1>";
-      }
-    </script>
-</body>
-</html>
